- **WriteGear:** 
    * Add support for GStreamer pipeline in WriteGear API's Non-Compression mode:
        + Implemented GStreamer Pipeline Mode to accept GStreamer pipeline as string to its output_filename parameter.
        + Added new special `-gst_pipeline_mode` attribute for its output_params parameter.
        + This feature provides flexible way to directly write video frames into GStreamer Pipeline with controlled bitrate. 
        + Added new docs and updated existing docs with related changes.
    * Added new `-ffpreheaders` special attribute to WriteGear's options parameter:
        + This attribute is specifically required to set special FFmpeg parameters in Compression Mode that are present at the starting of command(such as `-re`).
        + This attribute only accepts **list** datatype as value.
        + Added related docs.
- **NetGear:** 
    * Added bidirectional data transfer support by extending Bidirectional mode support to exclusive Multi-Clients and Multi-Servers modes:
        + Users will now able to send data bidirectionally in both Multi-Clients and Multi-Servers exclusive modes.
        + Bidirectional mode will no longer disables automatically when Multi-Clients and Multi-Servers modes already enabled.
        + Added new docs and updated existing docs with related changes.
- **Maintenance:** 
    * ***Added official support for **Python-3.10** legacies.***
    * Added `float` value support to `THREAD_TIMEOUT` optional parameter.
    * Added info about dropped support for Python-3.6 legacies through announcement bar.
    * Added `config.md` file for Issue templates.
    * Added title to Issue templates.
- **Docs:**
    * Added new Code Annotations
    * Added new icons to headings.
    * Added Advanced VideoGear usage example with CamGear backend.

- Setup.py:
    * Dropped support for Python-3.6 and below legacies.
    * Updated logging formatting.
    * Updated python_requires to `>=3.7`.
    * Bumped version to `0.2.5`.
- Helper:
    * Vidgear will now report current version on every run.
- Docs: 
    * Updated SSH tunneling docs context.
    * Excluded `docs` directory from CI envs.
    * Updated Zenodo badge and BibTeX entry.
    * Updated dark theme hue to `260`.
    * Updated Admonitions.
    * Additional warnings against pushing PR against VidGear's `testing` branch only.
    * Updated code comments.
- CI:
    * Removed support for Python-3.6 legacies from all workflows.
    * Updated NetGear's Exclusive Mode tests.
    * Added GStreamer Pipeline Mode tests.
- Maintenance: 
    * Updated Issue and PR templates.
    * Updated metadata.


- **⚠️  Dropped support for Python-3.6 legacies from vidgear.**

- NetGear: Fixed bidirectional mode overriding multi-clients mode's data.
- WriteGear: 
    * Fixed wrongly defined ffmpeg_preheaders.
    * Fixed condition logic bugs.
    * Fixed UnboundLocalError bug.
- Setup: Fixed uvicorn and aiortc dropped support for Python-3.6 legacies.
- CI: 
    * Fixed GitHub Actions interprets 3.10 as 3.1 if used without strings.
    * Fixed naming error in azure YAML.
- Docs:
    * Fixed codecov badge URL in README.md
    * Fixed hyperlinks in README.
    * Fixed indentation and spacing.
    * Fixed typos and updated context.
    * Removed dead code.
- Maintenance: 
    * Removed depreciated condition checks.

+++ vidgear/gears/camgear.py
@@ -22,7 +22,6 @@ limitations under the License.
 import cv2
 import time
 import queue
-import sys
 import logging as log
 from threading import Thread, Event
 
@@ -320,7 +319,7 @@ class CamGear:
         self.__thread_timeout = options.pop("THREAD_TIMEOUT", None)
         if self.__thread_timeout and isinstance(self.__thread_timeout, (int, float)):
             # set values
-            self.__thread_timeout = int(self.__thread_timeout)
+            self.__thread_timeout = float(self.__thread_timeout)
         else:
             # defaults to 5mins timeout
             self.__thread_timeout = None
@@ -329,7 +328,7 @@ class CamGear:
         # initialize queue for video files only
         if self.__threaded_queue_mode and isinstance(source, str):
             # define queue and assign it to global var
-            self.__queue = queue.Queue(maxsize=96)  # max len 96 to check overflow
+            self.__queue = queue.Queue(maxsize=96)  # max bufferlen 96 to check overflow
             # log it
             self.__logging and logger.debug(
                 "Enabling Threaded Queue Mode for the current video source!"
